MODULE 3

+++ Module 3.2/factorial.c
@@ -1,14 +1,14 @@
-#include<stdio.h>  
-int main()    
-{    
- int b=1;
- int a;    
- printf("Enter a number to find the factorial of: ");    
-  scanf("%d",&a);    
-  for(int i=1;i<=a;i++)
-  {    
-      b=b*i;    
-  }    
-  printf("Factorial of %d is: %d",a,b);    
-return 0;  
-}   
+#include <stdio.h>
+int main()
+{
+  long double b = 1;
+  int a;
+  printf("Enter a number to find the factorial of: ");
+  scanf("%d", &a);
+  for (int i = 1; i <= a; i++)
+  {
+    b = b * i;
+  }
+  printf("Factorial of %d is: %d", a, b);
+  return 0;
+}
